COMPLETE TASKS   -Se corrigieron los errores relacionados a eliminar partidas contables   -Se agregaron validaciones para evitar errores mas adelante.

+++ src/entries/entries.service.ts
@@ -866,6 +866,16 @@ export class EntriesService {
     return new ResponseSingleDTO(plainToClass(AccountingEntry, entrie));
   }
 
+  /**
+   *
+   * @param company ----Compañia con la que esta logada el usuario que invoica la peticion
+   * @param header  --Objeto de datos que se necesecitan para hacer el insert o update de la partida contable que se desea crear o actualizar
+   * @param details --Arreglo de objetos que poseen los datos necesarios o requeridos para crear los deralles (EntryDetailsDTO)
+   * @param type  --Tipo de acción que se desea realizar, ("create" || "update")
+   * @param id  --Parametro opcional, id de la partida contable que se desea actulizar
+   * @returns  -Retora el mensaje de exito en caso que asi sea acompañado del id del partida creada, si no retorna el mensaje de error
+   * --Metodo utilizado para estructurar la logica a realizar  ala hora de crear o actulizar partidas contables
+   */
   async createUpdateEntry(
     company: Company,
     header: any,
@@ -904,7 +914,9 @@ export class EntriesService {
         id: entry.id,
       };
 
-      await this.accountingEntryDetailRepository.deleteEntryDetail(entry.accountingEntryDetails.map((e) => e.id));
+      if (entry.accountingEntryDetails.length > 0) {
+        await this.accountingEntryDetailRepository.deleteEntryDetail(entry.accountingEntryDetails.map((e) => e.id));
+      }
     }
 
     const entryHeader = await this.accountingEntryRepository.createUpdateEntry(headerInsert, type);
@@ -936,21 +948,33 @@ export class EntriesService {
     };
   }
 
+  /**
+   *
+   * @param company --Compañia con la que esta logada el usuario que invoica la peticion
+   * @param id --Id de la partida contable que se desea eliminar
+   * @returns  Retorna un mensaje de exito o mensaje de error en caso que no  haya sido posible eliminar el registro
+   */
   async deleteEntry(company: Company, id: string): Promise<ResponseMinimalDTO> {
     const entry = await this.accountingEntryRepository.getEntry(company, id);
 
-    const resultDetails = await this.accountingEntryDetailRepository.deleteEntryDetail(
-      entry.accountingEntryDetails.map((e) => {
-        return { id: e.id, company };
-      }),
-    );
+    if (entry.accountingEntryDetails.length > 0) {
+      const resultDetails = await this.accountingEntryDetailRepository.deleteEntryDetail(
+        entry.accountingEntryDetails.map((e) => {
+          return { id: e.id, company };
+        }),
+      );
 
+      if (!resultDetails) {
+        throw new BadRequestException('No se ha podido eliminar la partida contable, contacta con tu administrador.');
+      }
+    }
     const resultHeader = await this.accountingEntryRepository.deleteEntry(company, id);
+
+    if (!resultHeader) {
+      throw new BadRequestException('No se ha podido eliminar la partida contable, contacta con tu administrador.');
+    }
     return {
-      message:
-        resultDetails && resultHeader
-          ? 'Se ha eliminado la partida contable correctamente'
-          : 'No se ha podido eliminar la partida contable',
+      message: 'Se ha eliminado la partida contable correctamente',
     };
   }
